feat: Rename `collections` extension to `collections.list`
