+++ hugr-core/src/std_extensions.rs
@@ -18,7 +18,7 @@ pub fn std_reg() -> ExtensionRegistry {
         arithmetic::conversions::EXTENSION.to_owned(),
         arithmetic::float_ops::EXTENSION.to_owned(),
         arithmetic::float_types::EXTENSION.to_owned(),
-        collections::EXTENSION.to_owned(),
+        collections::list::EXTENSION.to_owned(),
         logic::EXTENSION.to_owned(),
         ptr::EXTENSION.to_owned(),
     ]);
