+++ hugr-llvm/src/emit/test.rs
@@ -153,7 +153,7 @@ impl SimpleHugrConfig {
                         float_ops::EXTENSION_ID,
                         conversions::EXTENSION_ID,
                         logic::EXTENSION_ID,
-                        collections::EXTENSION_ID,
+                        collections::list::EXTENSION_ID,
                     ]),
                 ),
             )
