+++ hugr-py/src/hugr/std/collections/__init__.py
@@ -1,0 +1,1 @@
+"""Standard extensions for collection types and operations."""
